Move add story button to the right
The "add story" button has been moved to the right of the screen
with the column-toggle buttons. I think the UX is better like this.
The story is added below the button and we don't have to move the
cursor from left to right across the screen to fill the fields
like before.